node1-scanner/requirements.txt
# Async core
aiohttp>=3.9.0
httpx[http2]>=0.27.0
websockets>=13.0

# Data
msgpack>=1.0.8
redis[asyncio]>=5.0.0
supabase>=2.4.0
asyncpg>=0.29.0
orjson>=3.10.0

# ML / NLP (local, no API cost)
scikit-learn>=1.4.0
sumy>=0.11.0
nltk>=3.8.0
pandas>=2.2.0

# Exchange SDKs
python-binance>=1.0.19
pybit>=5.6.0
ccxt>=4.3.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Utils
python-dotenv>=1.0.1
pydantic>=2.7.0
structlog>=24.2.0
tenacity>=8.3.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-benchmark>=4.0.0
